backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
bcrypt==4.2.0
